requirements.txt
requests>=2.28.0
beautifulsoup4>=4.13.0
python-telegram-bot>=22.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
schedule>=1.2.0
pytest>=7.0.0
pytest-asyncio>=0.21.0 